DifficultyTransitions: interface and implementations for difficulty transitions
Extract logic from `BitcoinNetworkParameters` and `TestNet3Params`.

The old `BitcoinNetworkParams.isDifficultyTransitionPoint()` is kept
around as deprecated.
