+++ core/src/main/java/org/bitcoinj/params/TestNet3Params.java
@@ -20,12 +20,7 @@ package org.bitcoinj.params;
 import org.bitcoinj.base.BitcoinNetwork;
 import org.bitcoinj.base.Difficulty;
 import org.bitcoinj.core.Block;
-import org.bitcoinj.core.NetworkParameters;
 import org.bitcoinj.base.Sha256Hash;
-import org.bitcoinj.core.StoredBlock;
-import org.bitcoinj.core.VerificationException;
-import org.bitcoinj.store.BlockStore;
-import org.bitcoinj.store.BlockStoreException;
 
 import java.time.Instant;
 
@@ -42,8 +37,6 @@ public class TestNet3Params extends BitcoinNetworkParams {
     private static final Instant GENESIS_TIME = Instant.ofEpochSecond(1296688602);
     private static final long GENESIS_NONCE = 414098458;
     private static final Sha256Hash GENESIS_HASH = Sha256Hash.wrap("000000000933ea01ad0ee984209779baaec3ced90fa3f408719526f8d77f4943");
-    /** Spacing for the 20-minute difficulty exception. */
-    private static final int TESTNET_DIFFICULTY_EXCEPTION_SPACING = NetworkParameters.TARGET_SPACING * 2;
 
     public TestNet3Params() {
         super(BitcoinNetwork.TESTNET);
@@ -94,52 +87,5 @@ public class TestNet3Params extends BitcoinNetworkParams {
             }
         }
         return genesisBlock;
-    }
-
-    // February 16th 2012
-    private static final Instant testnetDiffDate = Instant.ofEpochMilli(1329264000000L);
-
-    @Override
-    public void checkDifficultyTransitions(final StoredBlock storedPrev, final Block nextBlock,
-        final BlockStore blockStore) throws VerificationException, BlockStoreException {
-        if (!isDifficultyTransitionPoint(storedPrev.getHeight()) && nextBlock.time().isAfter(testnetDiffDate)) {
-            // After 15th February 2012 the rules on the testnet change to avoid people running up the difficulty
-            // and then leaving, making it too hard to mine a block. On non-difficulty transition points, easy
-            // blocks are allowed if there has been a span of 20 minutes without one.
-            long timeDelta = nextBlock.time().getEpochSecond() - storedPrev.getHeader().time().getEpochSecond();
-            boolean isMinDiffBlock = nextBlock.difficultyTarget().equals(maxTarget());
-            if (timeDelta < 0 && isMinDiffBlock) {
-                // There is an integer underflow bug in Bitcoin Core that means mindiff blocks are accepted when time
-                // goes backwards. Thus, skip any further checks.
-                return;
-            } else if (timeDelta > TESTNET_DIFFICULTY_EXCEPTION_SPACING){
-                // 20 minute exception
-                checkDifficultyTarget(nextBlock, maxTarget());
-            } else {
-                // If no special rule applies, expect the last non-mindiff difficulty.
-                checkDifficultyTarget(nextBlock, backwardsSkipMindiffBlocks(storedPrev, blockStore).difficultyTarget());
-            }
-        } else {
-            super.checkDifficultyTransitions(storedPrev, nextBlock, blockStore);
-        }
-    }
-
-    private void checkDifficultyTarget(Block nextBlock, Difficulty expectedTarget) {
-        Difficulty newTarget = nextBlock.difficultyTarget();
-        if (!newTarget.equals(expectedTarget))
-            throw new VerificationException("Testnet block transition that is not allowed: " +
-                    expectedTarget + " vs " +
-                    newTarget);
-    }
-
-    private Block backwardsSkipMindiffBlocks(StoredBlock prev, BlockStore blockStore) throws BlockStoreException {
-        // Walk backwards until we find a block that doesn't have the easiest proof of work.
-        int interval = getInterval();
-        Difficulty maxTarget = maxTarget();
-        while (!prev.getHeader().equals(getGenesisBlock()) &&
-                prev.getHeight() % interval != 0 &&
-                prev.getHeader().difficultyTarget().equals(maxTarget))
-            prev = prev.getPrev(blockStore);
-        return prev.getHeader();
     }
 }
