+++ core/src/test/java/org/bitcoinj/core/StandardDifficultyTransitionsTest.java
@@ -1,0 +1,34 @@
+/*
+ * Copyright by the original author or authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.bitcoinj.core;
+
+import org.bitcoinj.base.BitcoinNetwork;
+import org.junit.Test;
+
+import static org.junit.Assert.assertFalse;
+import static org.junit.Assert.assertTrue;
+
+public class StandardDifficultyTransitionsTest {
+    private static final DifficultyTransitions TRANSITIONS = DifficultyTransitions.of(BitcoinNetwork.MAINNET);
+
+    @Test
+    public void isDifficultyTransitionPoint() {
+        assertFalse(TRANSITIONS.isDifficultyTransitionPoint(2014));
+        assertTrue(TRANSITIONS.isDifficultyTransitionPoint(2015));
+        assertFalse(TRANSITIONS.isDifficultyTransitionPoint(2016));
+    }
+}
