+++ core/src/test/java/org/bitcoinj/params/BitcoinNetworkParamsTest.java
@@ -34,13 +34,6 @@ public class BitcoinNetworkParamsTest {
     };
 
     @Test
-    public void isDifficultyTransitionPoint() {
-        assertFalse(BITCOIN_PARAMS.isDifficultyTransitionPoint(2014));
-        assertTrue(BITCOIN_PARAMS.isDifficultyTransitionPoint(2015));
-        assertFalse(BITCOIN_PARAMS.isDifficultyTransitionPoint(2016));
-    }
-
-    @Test
     public void isRewardHalvingPoint() {
         assertTrue(BITCOIN_PARAMS.isRewardHalvingPoint(209999));
